Deccrypt issues
Fixing issues with a PDF that decided to use literal strings for the ID field instead of the traditional hex values.
Also the same PDF used a 4-byte positive number instead of a negative number for the value of P in the encryption dictionay. Why can't people just follow the specification?